pyproject.toml
[project]
name = "aiobot"
version = "0.1.0"
description = "AI agent for weather-based activity recommendations and event suggestions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "gradio>=4.0.0",
]


[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (import sorting)
    "B",    # flake8-bugbear (common bugs)
    "UP",   # pyupgrade (modern Python)
    "D",    # pydocstyle (docstrings)
]

ignore = [
    "E501",  # line too long
    "D100",  # missing docstring in public module
    "D104",  # missing docstring in public package
    "D203",  # incorrect-blank-line-before-class
    "D213",  # multi-line-summary-second-line
]
